refactor: improve project structure and naming conventions
