+++ src/components/ui/ButtonNavigation/index.tsx
@@ -1,9 +1,12 @@
 import styles from './ButtonNavigation.module.css';
-interface ButtonNavProps {
+interface ButtonNavigationProps {
   text: string;
   onClick?: () => void;
 }
-export default function ButtonNav({ text, onClick }: ButtonNavProps) {
+export default function ButtonNavigation({
+  text,
+  onClick,
+}: ButtonNavigationProps) {
   return (
     <button className={styles.button_nav} type="button" onClick={onClick}>
       {text}
