+++ src/components/ui/icons/MinusIcon/index.tsx
