+++ src/pages/CartPage/index.tsx
@@ -1,4 +1,4 @@
-import ShoppingCartComponent from '@/components/ShoppingCartComponent';
+import ShoppingCartComponent from '@/components/ShoppingCart';
 
 export default function CartPage() {
   return <ShoppingCartComponent />;
