+++ src/pages/CategoriesPage/index.tsx
@@ -1,8 +1,4 @@
 import AllCategories from '@/components/AllCategories';
 export default function CategoriesPage() {
-  return (
-    <div>
-      <AllCategories />
-    </div>
-  );
+  return <AllCategories />;
 }
