+++ src/pages/ProductsByCategoryPage/index.tsx
@@ -1,4 +1,4 @@
-import ProductListPage from '@/components/ProductsListComponent';
+import ProductListPage from '@/components/ProductsList';
 
 export default function ProductsByCategoryPage() {
   return <ProductListPage />;
