+++ src/pages/SingleProductPage/index.tsx
@@ -1,4 +1,4 @@
-import ProductInfoComponent from '@/components/ProductInfoComponent';
+import ProductInfoComponent from '@/components/ProductInfo';
 
 export default function SingleProductPage() {
   return <ProductInfoComponent />;
